Switch to use System.nanoTime in measureTimeMillis
Previously measureTimeMillis used System.currentTimeMillis which is not
ideal for several reasons. Firstly it often has a resolution of 10s of
milliseconds and secondly because it can be adjusted by something like
an NTP update so the returned value could be negative.

Instead it now delegates to measureNanoTime which uses System.nanoTime
which has the highest resolution the system provides as well as being
guaranteed to be monotonically increasing.

This fixes KT-52686 and KT-24431