[Gradle][MPP] Flatten transformed metadata library directories
The main purpose of this commit is to fix KT-48135 and to
de-duplicate libraries for the IDE whilst also minimizing
zip operations during IDE import.

However, with this commit the metadata library directories layout
also gets changed for CLI builds. The overall layout
is now more shared with cinterops and the IDE.

^KT-48135 Verification Pending