FIR flexible type: add ConeFlexibleTypeWithEnhancement
The existing code sets a flexible type with an annotation as a simple
type (ConeSimpleKotlinType) by applying the enhancement. Since most of
methods of platform types e.g., `java.lang.String.valueOf(1)` has an
annotation e.g., `NotNull`, the logic makes them as a simple type with
an annotation. However, it must be a flexible type e.g.,
`String..String?`. Making it a simple type can cause side effects for
users of analysis APIs because they determine a platform type based on
whether it is a flexible type or not.

This commit adds `ConeFlexibleTypeWithEnhancement` to match the
counterpart FE1.0 `FlexibleTypeWithEnhancement`, so that a flexible type
with an annotation can be `ConeFlexibleTypeWithEnhancement` after the
enhancement.