[Gradle][MPP] Include artifact checksum string in transformed metadata klib file names
This can be used for locally published snapshot libraries.
Without it, the import will see that there is already
a library with same coordinates in .gradle/kotlin/... and
therefore skip extracting it.

Including the 32 checksum at the end of the file will ensure,
that if a different artifact is the source of the metadata,
then a new file will be created.

This approach is generally more desirable than cleaning the
.gradle/kotlin directory to avoid the IDE of re-indexing
after import.

^KT-48135 Verification Pending

(cherry picked from commit 75cc378cac2b780104cb6c01c7ccf53e72bc50c1)