Allow lambda memoization inside composables returning a value (again)
Reverts previous revert of this commit (aosp/2215643) with additional check to ensure remembered param is not inlined or crossinlined lambda parameter of the inline function.
The edge case reported in b/246399235 ( https://issuetracker.google.com/issues/246399235 ) also applies to functions returning no values, this commit adds additional test for it as well.

Test: LambdaMemoizationTransformTests
Change-Id: I60acf8270e3979bc08bc073fde375efc371c4757 ( https://android-review.git.corp.google.com/q/I60acf8270e3979bc08bc073fde375efc371c4757 )