FIR: create substitution overrides for static Java members
Because static members have no receivers, the class through which the
member is accessed has to be encoded in the declaration itself, and this
seems like the most logical way to do this. And the class *needs* to be
encoded because the supertype may be invisible, which causes false
positive INVISIBLE_REFERENCE errors; even if those errors are ignored,
a reference to the super class in the bytecode will cause a JVM access
violation (whereas accessing the same member through the subclass is
perfectly fine).

^KT-53441 Fixed