Skip shortening for class with type parameter from parent
If a class has an implicit type parameter that is supposed to be
provided to its parent, we cannot shorten the parent, because shortening
it will result in losing the type parameter. For example:

class Outer<T> (inner: Outer<T><caret>.Inner) {  // we cannot shorten it.
  class Inner
}

This commit lets the shortener skip this case.