FIR ReferenceShortener: handle type parameter without PSI
In the following example, the built-in type ArrayList has a type
parameter for itself (not its parent), but it does not have PSI.

```
import java.util.ArrayList

fun test() {
    val a = java.<caret>util.ArrayList<Int>()
}
```

The existing type parameter check for shortening a type determines we
cannot shorten the above example because ArrayList has a type parameter
whose containing class PSI is different from the PSI of ArrayList.

It is not true because just the class containing the type parameter does
not have PSI (null). It is a built-in type. This commit replaces the PSI
comparison with the symbol comparison between the class symbol and the
class symbol containing the type parameter.