Restore kludge to deal with typed null values in spilled var analysis
This is a partial revert of 0fc676a20cf (and thus a re-application of
7579be6c685). The main problem in that commit was the change in
`OptimizationBasicInterpreter.binaryOperation` where some AALOAD
operations started to produce null values, which became an issue since
65c153a722b where casts on such values started to be recognized as
unnecessary and became incorrectly optimized out.

Returning `REFERENCE_VALUE` for any AALOAD would seem like the most
natural solution, but unfortunately it breaks the test added in
7579be6c685 because the null value gets spilled again and coroutine
transformer can't deduce its type, which leads to VerifyError.

So in this change, I'm restoring the original implementation from
7579be6c685 as it seems like the safest option.

 #KT-54802 Fixed