Drop outdated KDoc and fix findSourceFirDeclarationByExpression signature
Review: https://jetbrains.team/p/kt/reviews/7690

Make findSourceFirDeclarationByExpression signature more specific.
`KtDeclaration` is an inheritor of `KtExpression`

`KtLambdaExpression` was changed to `KtFunctionLiteral` in
c24ad0ba51888df0aff8a72ee09a386049c760dd. I run analysis-api-fir tests
and ensured that `KtFunctionLiteral` appears in runtime instead of
`KtFunctionLiteral`

So `ktDeclaration` should be either `KtDeclaration` or
`KtFunctionLiteral`. But we don't need a KDoc for that because
`KtFunctionLiteral` is an inheritor of `KtDeclaration`, so type-system
already checks this invariant