Fix configurations in dependency handler of KotlinWithJavaCompilation
Implementations of compilation and its dependency holder used to return
different configuration names as HasKotlinDependencies implementation.
Compilation setup creates configuration with the names, defined in the
compilation itself, but when in a project dependencies are added via the
dependency handler of the compilation, dependencies holder with its own
names is used. The attempt to use missing configuration causes failure.

The fix removes duplication, only the dependencies holder now defines
configuration names and KotlinWithJavaCompilation reuses the logic of
AbstractKotlinCompilation and delegates it to the holder.

KT-55019