[FIR] Properly report about recursive implicit types
Before the change `InferenceError` was also reported, but it was never
collected. It seems that in may be safely replaced with
`RecursionInImplicitTypes` for full body resolve, and the only place
where makes no sense is the contracts resolution.

A custom returnTypeCalculator is introduced for contracts to preserve
the existing behavior. The fact that `InferenceError` is not reported
here is a separate story.

Here is a test to play with:
FirDiagnosticTestSpecGenerated.NotLinked.Contracts.Declarations
.ContractBuilder.Common.Neg.test12