[FIR] Add one more handy test
This test checks both the type mismatch diagnostics from checkers and
the ones from inference (collected from error named references). It also
ensures there is a symmetry in the way diagnostics are
reported via the checkers and via the error named refs.

I added this test retroactively, because it turned out to be an easier
way to observe the changes.