[FIR] Fix missing RETURN_TYPE_MISMATCH-s
`returnTargetIfFromReturnType` is not the same as the
`expectedTypeMismatchIsReportedInChecker` flag:
the return target is needed for the diagnostic.

In places where `RETURN_TYPE_MISMATCH` stopped reporting
it stopped, because the checker now ignores `FirResolvable`s with
an `FirErrorNamedReference`, and since it is an error named reference,
it will be collected via `toFirDiagnostics`, which means there will be
some sort of error happening here.

In places where new `RETURN_TYPE_MISMATCH` was added, it is reported
from within `toFirDiagnostics`, because the flag `*reportedInChecker`
is not set anymore. This diagnostic is quite often reported for
arguments of calls (e.g. synthetic calls for `when`s/`if`s). These are
examples of cases when we had missing diagnostics that the checker
failed to detect