[FIR] Properly report INITIALIZER_TYPE_MISMATCH
As before, in cases where one `INITIALIZER_TYPE_MISMATCH`
was removed it was removed because the checker now
does not check `FirResolvable`s with error named references.

`TYPE_MISMATCH` was replaced by `INITIALIZER_TYPE_MISMATCH`, because
`ConstraintSystemError.toDiagnostic` now selects
the proper diagnostic.

`TYPE_MISMATCH`/`INITIALIZER_TYPE_MISMATCH` may be omitted if one of the
types is an error type - in this case there will be
a proper diagnostic reported for that.