[FIR] Fix the missing INFERENCE_ERROR
In a bunch of cases it seems redundant, but on the other hand there are
cases then it's just missed. This change fixes those tests so
that later we will not lose it.

The `12.fir.kt` is ignored, because FirIdeSpecTest fails if the
`.fir.kt` file differs from the `.kt` file. This commit
is not about fixing anything, but rather reporting the
existing errors.