K2: Fix false-positive type equality for different captured types
This is mostly a revert of 2f61a2f56f55a55c8a11efde94bc5984418a8cab
There, we erroneously assumed that we may take captured types as equal
if they are based on the same-typed projections.

Each instance of capturing defines its own captured type,
that should not be equal to any other type captured in other place.

The most obvious consequence is the change in addAllProjection.fir.kt:
one cannot use an instance as an argument when expected type
is captured type based on the same instance.

Otherwise, it would lead to CCE if we allowed to put arbitrary charsequences
to the list that initially was a MutableList<String>

All other test data changes are irrelevant and will be fixed
in the subsequent commits