Cache daemon connection & compiler classloader using shared service
Previously:
* KGP created a new connection to the daemon for each compile task.
Reconnecting to the daemon is relatively fast, but it might add up
to a few seconds, when there are hundreds of tasks;
* KGP created a new isolated classloader for each compile task,
when the in-process execution strategy was used. This made
the in-process mode basically useless, because the compiler classes
were recompiled by the JIT for each invocation of the compiler.

This change introduces a new shared service
to cache both for a duration of one build.
Caching between builds might be even better,
but Gradle Build services are shutdown at the end of each build
https://github.com/gradle/gradle/issues/20867