Fix KT-50777
The problem is that the last valid start-index of a substring is the
length of the string, not the last valid index of the string. This is
only a problem, when searching for the empty string. The fix is to start
any search for a last occurence of a substring from index=length instead
of index=lastIndex.