Add isInline to the Group class of ui tooling data
The `Group` class in the ui tooling data API now has a field
that indicates if the group is for an inline function.

Relnote: """the ui tooling data class `Group` now has a field,
`isInline`, that indicates if the group is for a call to an
inline composable function. If `isInline` is `true` then the
call is to an inline composable function. However, the value
might be false for calls to inline composable functions that
are from modules that are compiled with a version of the
compose compiler plugin that doesn't generate the inline
function information."""

Test: ./gradlew :compose:compiler:c-h:i-t:tDUT :compose:ui:ui-tooling-data:cC
Fixes: 245403738 ( https://issuetracker.google.com/issues/245403738 )
Change-Id: Idb84635d4b427de2c4636792307f7e60360a9f0a ( https://android-review.googlesource.com/q/Idb84635d4b427de2c4636792307f7e60360a9f0a )

Moved from: https://github.com/androidx/androidx/commit/d0154f99758da04183030de9521d62b31013ff8f