[FIR] KT-55411: Remove the VariableAccess check
This check was needed for java fields like in this
test: staticsFromJavaNew.kt.

The actual situation is the following:
There is a `givenExtensionReceiverOptions.isEmpty()` check in
`MemberScopeTowerLevel.processMembers` that prevents
searching for synthetics if `extensionReceiver != null`. This means,
when we can't remember if a receiver resulted in empty scopes
(implicitReceiverValuesWithEmptyScopes)
because we may first check it with
`extensionReceiver = implicitReceiver`, thus not getting anything.
So, `processLevel()` for member
scope from receivers with `extensionReceiver = null` yields all
the same results as `extensionReceiver = receiver`
plus some more.

The new change allows to not check the same scopes in case of
VariableAccess, because when we check the first time,
we already check with the synthetics in mind.