[FIR] KT-55411: Remove the VariableAccess check
This check was needed for java fields like in this
test: staticsFromJavaNew.kt.

The actual situation is the following:
There is a `givenExtensionReceiverOptions.isEmpty()` check in
`MemberScopeTowerLevel.processMembers` that prevents
searching for synthetics if `extensionReceiver != null`. This means,
when we can't remember if a receiver resulted in empty scopes
(implicitReceiverValuesWithEmptyScopes)
because we may first check it with
`extensionReceiver = implicitReceiver`, thus not getting anything.
So, `processLevel()` for member
scope from receivers with `extensionReceiver = null` yields all
the same results as `extensionReceiver = receiver`
plus some more.

The new change allows to not check the same scopes in case of
VariableAccess, because when we check the first time,
we already check with the synthetics in mind.

The solution below relies on the fact that synthetic properties
are not checked for `hasConsistentExtensionReceiver`
(this check happens inside `scope.collectCandidates`) call, but
not for synthetics.

One could argue that the hypothetical proper solution could be:

- also check `hasConsistentExtensionReceiver` for synthetics
- if it's false, `return ProcessResult.FOUND`
  - this would prevent caching, and wait until we check
    the member scope without a receiver

But since java fields can't have extension receivers we may never get
a receiver mismatch when accessing it via any dispatch receiver.
Type mismatch doesn't happen with expression receivers either, because
in these cases they may only be found when we look in the
member scope of the explicit receiver
(see: `runResolverForExpressionReceiver`). When enumerating the tower
levels after it, we don't search for synthetics, because
we get `typeForSyntheticScope = null` in such cases (because the flag
`givenExtensionReceiverCameFromImplicitReceiver`
is not being set)