[Gradle] Deprecate scoped metadata configurations
They are unnecessary and `resolvableMetadataConfiguration` is enough.
Since we need metadata to be used only for Compilation scope. And in
IDE import there is no need for scoped transformations either.
Therefore, we can keep only user-facing configurations separated by
scope (api, implementation, compileOnly). And use them as inputs for
`resolvableMetadataConfiguration`.

^KT-55312 Fixed