Make more IME classes and constants stable
This CL makes KeyboardActions, KeyboardOptions, and all of
KeyboardOptions’ transitive properties stable, as well as all of the
constant instances of these classes.

This lets us treat these expressions as static (a nice win since they
appear as default arguments for several Composables). The least trivial
addition of `@Stable` is in `KeyboardActions`, which now has an
implementation of `equals` and `hashCode`. This isn’t strictly necessary
for us to treat instances as static, but it should be a nice win for
consumers that specify one set of keyboard actions that doesn’t change.

Bug: 184876305 ( https://issuetracker.google.com/issues/184876305 )
Test: ./gradlew :compose:compiler:c-h:i-t:tDUT
Change-Id: I844466c112f90ca850b6cd40a76f56186310f0c1 ( https://android-review.git.corp.google.com/q/I844466c112f90ca850b6cd40a76f56186310f0c1 )
Relnote: N/A