Detect static immutable class constructor calls
Currently, our checks for static function calls are a bit basic around
constructor calls. A constructor is only ever a candidate to be static
if it’s a value class.

This CL fleshes out the existing logic to check other constructor
invocations to see if they’re static. The criteria for a static
constructor call is that the class must be `@Immutable`, and all of the
arguments passed into the constructor must be static expressions.

Test: ./gradlew :compose:compiler:c-h:i-t:tDUT
Change-Id: Ib06df2f5fcf75cf10c05f6e367ff1c22e88abd31 ( https://android-review.git.corp.google.com/q/Ib06df2f5fcf75cf10c05f6e367ff1c22e88abd31 )