[FIR] Remove false caches
Those sets are never hit which means that we never revisit the same
thing twice. This was determined by calculating the statistics of
what is actually called and what is not.

The code is written in such a way that it's easy to reintroduce some
caches if we decide we want need them.