Introduce Preview language version
This lv is needed to stabilize IC tests, that rely on incompatible metadata, which the compiler can produce if it has an opportunity to set language version with {minor+2} version from the current. Occasionally such a language version that satisfies the requirements exists not every release, so IC logic can be broken and this will be missed by testing. But the hidden problem will appear again when lv {major, minor+2} will be added again in future releases. To prevent such cases Preview language version is introduced. This lv will not be suggested to the user during hints, but it will be possible to set it manually.

(#KT-54215, #KTIJ-20954) Relates